Refactoring Optimizers struct to optimizers module
