+++ delta/src/neuralnet/layers/conv_1d.rs
@@ -176,7 +176,7 @@ impl Layer for Conv1D {
     /// * `optimizer` - The optimizer to use.
     fn update_weights(
         &mut self,
-        optimizer: &mut Box<dyn crate::common::optimizer::Optimizer>,
+        optimizer: &mut Box<dyn crate::optimizers::Optimizer>,
     ) -> Result<(), LayerError> {
         if !self.trainable {
             return Ok(());
