+++ delta/src/neuralnet/layers/dense.rs
@@ -27,10 +27,11 @@
 //! OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 //! OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use crate::common::{Layer, LayerError, Optimizer, Shape, Tensor};
+use crate::common::{Layer, LayerError, Shape, Tensor};
 use log::debug;
 use serde_json;
 use crate::activations::Activation;
+use crate::optimizers::Optimizer;
 
 /// A dense (fully connected) layer.
 #[derive(Debug)]
