+++ delta/src/optimizers/mod.rs
@@ -29,4 +29,24 @@
 
 pub mod adam;
 
+use std::fmt::Debug;
 pub use adam::Adam;
+use crate::common::{OptimizerError, Tensor};
+
+/// A trait representing an optimizer for training neural networks.
+pub trait Optimizer: Debug {
+    /// Performs an optimization step using the given weights and gradients.
+    ///
+    /// # Arguments
+    ///
+    /// * `weights` - A mutable reference to the weights tensor.
+    /// * `gradients` - A reference to the gradients tensor.
+    fn step(&mut self, weights: &mut Tensor, gradients: &Tensor) -> Result<(), OptimizerError>;
+}
+
+/// A struct representing the configuration for an optimizer.
+#[derive(Debug)]
+pub struct OptimizerConfig {
+    /// The learning rate for the optimizer.
+    pub learning_rate: f32,
+}
